+++ tensorflow_decision_forests/tensorflow/core_test.py
@@ -223,6 +223,20 @@ class SimpleMlTfTest(tf.test.TestCase):
         core.column_type_to_semantic(data_spec_pb2.ColumnType.CATEGORICAL_SET),
         core.Semantic.CATEGORICAL_SET)
 
+  def test_replace_port_in_address(self):
+    self.assertEqual(
+        core.replace_port_in_address("127.0.0.1:1234", 4567), "127.0.0.1:4567")
+    self.assertEqual(
+        core.replace_port_in_address("[::]:1234", 4567), "[::]:4567")
+    self.assertEqual(
+        core.replace_port_in_address("www.domain.com:1234", 4567),
+        "www.domain.com:4567")
+    self.assertEqual(
+        core.replace_port_in_address("127.0.0.1:1234", 4567), "127.0.0.1:4567")
+
+    with self.assertRaises(ValueError):
+      core.replace_port_in_address("127.0.0.1", 4567)
+
 
 if __name__ == "__main__":
   tf.test.main()